requirements.txt
Django==1.3.1
Jinja2==2.6
Markdown==2.1.1
MySQL-python==1.2.3
PIL==1.1.7
South==0.7.3
-e git+https://github.com/concentricsky/django-basic-models.git@9522b90cc6eafabb45671b40ff8558ee00b8202f#egg=django_basic_models-dev
-e git+https://github.com/concentricsky/django-cachemodel.git@b1f795ef07eff33ff57cd683fcab0ec968c251df#egg=django_cachemodel-dev
django-debug-toolbar==0.9.4
django-tastypie==0.9.11
jingo==0.4
mimeparse==0.1.3
python-dateutil==1.5
wsgiref==0.1.2
